# api-contract/openapi.yaml
openapi: 3.0.3
info:
  title: Terraform Cost Intelligence Platform API
  version: 2.0.0
  description: |
    **CANONICAL API CONTRACT** - Single source of truth for all platform interfaces.
    
    ## Contract Discipline
    
    - All services MUST conform to this specification
    - All responses MUST use the ApiResponse envelope
    - All job statuses MUST use the canonical JobStatus enum
    - No undocumented endpoints allowed
    - No custom response shapes allowed
    
    ## Response Envelope
    
    Every endpoint returns:
    ```json
    {
      "success": boolean,
      "data": <payload or null>,
      "error": <Error object or null>,
      "correlation_id": "uuid"
    }
    ```

servers:
  - url: http://localhost:8000
    description: Local development
  - url: /api
    description: Production (behind reverse proxy)

components:
  schemas:
    ApiResponse:
      $ref: './schemas/ApiResponse.yaml#/ApiResponse'
    
    Error:
      $ref: './schemas/Error.yaml#/Error'
    
    Job:
      $ref: './schemas/Job.yaml#/Job'
    
    JobStatus:
      $ref: './schemas/JobStatus.yaml#/JobStatus'
    
    UsageProfile:
      $ref: './schemas/UsageProfile.yaml#/UsageProfile'
    
    CostResult:
      $ref: './schemas/CostResult.yaml#/CostResult'
    
    ResourceCost:
      $ref: './schemas/CostResult.yaml#/ResourceCost'

  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

paths:
  /health:
    get:
      summary: Health check
      tags: [Health]
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "healthy"
                  timestamp:
                    type: string
                    format: date-time

  /api/usage-profiles:
    get:
      summary: Get available usage profiles
      tags: [Usage Profiles]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of usage profiles
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/UsageProfile'
        '503':
          description: Usage service unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  /api/usage-profiles/validate:
    post:
      summary: Validate usage profile configuration
      tags: [Usage Profiles]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - profile_id
              properties:
                profile_id:
                  type: string
                  example: "prod"
      responses:
        '200':
          description: Profile is valid
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          valid:
                            type: boolean
                            example: true
        '400':
          description: Invalid profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  /api/jobs:
    post:
      summary: Create a new cost estimation job
      tags: [Jobs]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - upload_id
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 255
                upload_id:
                  type: string
                  format: uuid
                usage_profile:
                  type: string
                  default: "prod"
      responses:
        '201':
          description: Job created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Job'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    
    get:
      summary: List jobs with pagination
      tags: [Jobs]
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/JobStatus'
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      responses:
        '200':
          description: Paginated list of jobs
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Job'
                      meta:
                        type: object
                        properties:
                          pagination:
                            type: object
                            properties:
                              page:
                                type: integer
                              page_size:
                                type: integer
                              total:
                                type: integer
                              total_pages:
                                type: integer

  /api/jobs/{job_id}:
    get:
      summary: Get job details
      tags: [Jobs]
      security:
        - BearerAuth: []
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Job details
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Job'
        '404':
          description: Job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    
    delete:
      summary: Delete a job
      tags: [Jobs]
      security:
        - BearerAuth: []
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Job deleted successfully
        '404':
          description: Job not found

  /api/jobs/{job_id}/results:
    get:
      summary: Get cost estimation results
      tags: [Jobs]
      security:
        - BearerAuth: []
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Cost estimation results
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/CostResult'
        '400':
          description: Job not completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          description: Job not found or results not available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  /api/jobs/{job_id}/events:
    get:
      summary: Get job status events (WebSocket upgrade)
      tags: [Jobs]
      security:
        - BearerAuth: []
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '101':
          description: Switching to WebSocket protocol
        '404':
          description: Job not found

tags:
  - name: Health
    description: Health check endpoints
  - name: Usage Profiles
    description: Usage profile management
  - name: Jobs
    description: Cost estimation job management